Fix for jasminb/jsonapi-converter#131
Added javadoc and and renamed exception.
